Put Double in argument

+++ src/Oct/Task23102024/Task1.java
@@ -12,7 +12,7 @@ public class Task1 {
     static int mul(int num1, int num2){
         return num1*num2;
     }
-    static double  div(int num1,int num2){
+    static double div(double num1,double num2){
         return num1/num2;
     }
